refactor: move out helper functions
