+++ @commitlint/load/package.json
@@ -2,8 +2,8 @@
   "name": "@commitlint/load",
   "version": "8.3.5",
   "description": "Load shared commitlint configuration",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
+  "main": "lib/load.js",
+  "types": "lib/load.d.ts",
   "files": [
     "lib/"
   ],
